support_bot/requirements.txt
langchain>=0.1.0
chromadb>=0.4.22
python-dotenv>=1.0.0
openai>=1.12.0
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
numpy>=2.1.0 